[arcane,utils] Déplace les méthodes du namespace 'Convert' de 'ValueConvert.h' vers 'Convert.h'.
