+++ arcane/src/arcane/utils/Convert.cc
@@ -23,6 +23,13 @@
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
+/*!
+ * \brief Fonctions pour convertir un type en un autre.
+ * \namespace Arcane::Convert
+ */
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
 
 namespace Arcane
 {
