+++ arcane/src/arcane/utils/ValueConvert.h
@@ -18,6 +18,7 @@
 #include "arcane/utils/String.h"
 #include "arcane/utils/Array.h"
 #include "arcane/utils/NumericTypes.h"
+#include "arcane/utils/Convert.h"
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
@@ -349,56 +350,6 @@ inline const char* typeToName(const Int64Array&)
 {
   return "Int64[]";
 }
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-
-namespace Convert
-{
-
-  template <typename T>
-  class Type;
-
-  template <typename T>
-  class ScalarType
-  {
-   public:
-
-    //! Convertit \a s en le type \a T
-    ARCANE_UTILS_EXPORT static std::optional<T> tryParse(StringView s);
-
-    /*!
-   * \brief Convertit \a s en le type \a T.
-   *
-   * Si \a s.empty() est vrai, alors retourne \a default_value.
-   */
-    static std::optional<T>
-    tryParseIfNotEmpty(StringView s, const T& default_value)
-    {
-      return (s.empty()) ? default_value : tryParse(s);
-    }
-
-    /*!
-   * \brief Convertit la valeur de la variable d'environnement \a s en le type \a T.
-   *
-   * Si platform::getEnvironmentVariable(s) est nul, return std::nullopt.
-   * Sinon, retourne cette valeur convertie en le type \a T. Si la conversion
-   * n'est pas possible, retourne std::nullopt si \a throw_if_invalid vaut \a false ou
-   * lève une exception s'il vaut \a true.
-   */
-    ARCANE_UTILS_EXPORT static std::optional<T>
-    tryParseFromEnvironment(StringView s, bool throw_if_invalid);
-  };
-
-  //! Spécialisation pour les types scalaires
-  template <> class Type<Int64> : public ScalarType<Int64>
-  {};
-  template <> class Type<Int32> : public ScalarType<Int32>
-  {};
-  template <> class Type<Real> : public ScalarType<Real>
-  {};
-
-} // namespace Convert
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
